feat: 선택한 이미지 타입 리턴하는 함수 추가 - 카메라로 찍은 이미지, 스크린샷, 기타를 지정해서 리턴하는 함수 - fetchGallery  Colum (앨범 아이디, 앨범 명) 추가 - impl 클래스에서 private 으로된 함수들 내부적 Extensions 클래스로 옮김 (impl 로직을 좀더 집중있게 보기 위함)
